activity-app: setup first screen app
