+++ app/src/main/java/joaquim/lop/io/bored/MyApplication.kt
@@ -1,0 +1,8 @@
+package joaquim.lop.io.bored
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MyApplication : Application() {
+}
