+++ app/src/main/java/joaquim/lop/io/bored/di/Module.kt
@@ -1,0 +1,56 @@
+package joaquim.lop.io.bored.di
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import joaquim.lop.io.bored.data.remote.ServiceApi
+import joaquim.lop.io.bored.util.Constants.BASE_URL
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object Module {
+
+    @Singleton
+    @Provides
+    fun provideOkHttpClient(): OkHttpClient {
+        val logging = HttpLoggingInterceptor()
+        logging.level = HttpLoggingInterceptor.Level.BODY
+
+        return OkHttpClient().newBuilder()
+            .addInterceptor { chain ->
+                val newUrl = chain.request().url
+                    .newBuilder()
+                    .build()
+
+                val newRequest = chain.request()
+                    .newBuilder()
+                    .url(newUrl)
+                    .build()
+                chain.proceed(newRequest)
+            }
+            .addInterceptor(logging)
+            .build()
+    }
+
+    @Singleton
+    @Provides
+    fun provideRetrofit(client: OkHttpClient): Retrofit {
+        return Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .client(client)
+            .build()
+    }
+
+    @Singleton
+    @Provides
+    fun provideServiceApi(retrofit: Retrofit): ServiceApi {
+        return retrofit.create(ServiceApi::class.java)
+    }
+}
