+++ app/src/main/java/joaquim/lop/io/bored/ui/activity/MainActivity.kt
@@ -1,0 +1,28 @@
+package joaquim.lop.io.bored.ui.activity
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.fragment.NavHostFragment
+import dagger.hilt.android.AndroidEntryPoint
+import joaquim.lop.io.bored.R
+import joaquim.lop.io.bored.databinding.ActivityMainBinding
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+
+    private lateinit var navHostFragment: NavHostFragment
+    private lateinit var binding: ActivityMainBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        initViews(binding)
+    }
+
+    private fun initViews(binding: ActivityMainBinding) {
+        navHostFragment =
+            supportFragmentManager.findFragmentById(R.id.fragmentContainerView) as NavHostFragment
+    }
+}
