+++ app/src/main/java/joaquim/lop/io/bored/ui/activity/SplashActivity.kt
@@ -1,0 +1,33 @@
+package joaquim.lop.io.bored.ui.activity
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import joaquim.lop.io.bored.R
+import joaquim.lop.io.bored.databinding.ActivitySplashBinding
+
+
+class SplashActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivitySplashBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_splash)
+        binding = ActivitySplashBinding.inflate(layoutInflater)
+        val view = binding.root
+        setContentView(view)
+        setupSplash()
+        supportActionBar?.hide()
+    }
+
+    private fun setupSplash() = with(binding) {
+        animationView.alpha = 0f
+        animationView.animate().setDuration(1000).alpha(1f).withEndAction {
+            val intent = Intent(this@SplashActivity, MainActivity::class.java)
+            startActivity(intent)
+            overridePendingTransition(android.R.anim.fade_in, android.R.anim.fade_out)
+            finish()
+        }
+    }
+}
