+++ app/src/main/java/joaquim/lop/io/bored/util/Constants.kt
@@ -1,0 +1,6 @@
+package joaquim.lop.io.bored.util
+
+object Constants {
+
+    const val BASE_URL = "http://www.boredapi.com/api/"
+}
