--- app/src/main/java/joaquim/lop/io/marvelappstarter/MyApplication.kt
@@ -1,8 +1,0 @@
-package joaquim.lop.io.marvelappstarter
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class MyApplication : Application() {
-}
