--- app/src/main/java/joaquim/lop/io/marvelappstarter/data/model/character/ActivityModelData.kt
@@ -1,9 +1,0 @@
-package joaquim.lop.io.marvelappstarter.data.model.character
-
-import com.google.gson.annotations.SerializedName
-import java.io.Serializable
-
-data class ActivityModelData(
-    @SerializedName("results")
-    val results: ActivityModel
-) : Serializable
