--- app/src/main/java/joaquim/lop/io/marvelappstarter/data/remote/ServiceApi.kt
@@ -1,11 +1,0 @@
-package joaquim.lop.io.marvelappstarter.data.remote
-
-import joaquim.lop.io.marvelappstarter.data.model.character.ActivityModelResponse
-import retrofit2.Response
-import retrofit2.http.GET
-
-interface ServiceApi {
-
-    @GET("activity")
-    suspend fun list(): Response<ActivityModelResponse>
-}
