--- app/src/main/java/joaquim/lop/io/marvelappstarter/repository/MarvelRepository.kt
@@ -1,10 +1,0 @@
-package joaquim.lop.io.marvelappstarter.repository
-
-import joaquim.lop.io.marvelappstarter.data.remote.ServiceApi
-import javax.inject.Inject
-
-class MarvelRepository @Inject constructor(
-    private val api: ServiceApi
-) {
-    suspend fun list() = api.list()
-}
