--- app/src/main/java/joaquim/lop/io/marvelappstarter/ui/activity/MainActivity.kt
@@ -1,35 +1,0 @@
-package joaquim.lop.io.marvelappstarter.ui.activity
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.setupWithNavController
-import dagger.hilt.android.AndroidEntryPoint
-import joaquim.lop.io.marvelappstarter.R
-import joaquim.lop.io.marvelappstarter.databinding.ActivityMainBinding
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var navHostFragment: NavHostFragment
-    private lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        initViews(binding)
-    }
-
-    private fun initViews(binding: ActivityMainBinding) {
-        navHostFragment =
-            supportFragmentManager.findFragmentById(R.id.fragmentContainerView) as NavHostFragment
-        val navController = navHostFragment.navController
-
-        binding.bottomNavigation.apply {
-            setupWithNavController(navController)
-            setOnNavigationItemReselectedListener { }
-        }
-    }
-}
