--- app/src/main/java/joaquim/lop/io/marvelappstarter/ui/base/BaseFragment.kt
@@ -1,32 +1,0 @@
-package joaquim.lop.io.marvelappstarter.ui.base
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModel
-import androidx.viewbinding.ViewBinding
-
-abstract class BaseFragment<VB : ViewBinding, VM : ViewModel> : Fragment() {
-
-    private var _binding: VB? = null
-    protected val binding get() = _binding!!
-    protected abstract val viewModel: VM
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = getViewBinding(inflater, container)
-        return binding.root
-    }
-
-    protected abstract fun getViewBinding(inflater: LayoutInflater, container: ViewGroup?): VB
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
