--- app/src/main/java/joaquim/lop/io/marvelappstarter/util/Constants.kt
@@ -1,6 +1,0 @@
-package joaquim.lop.io.marvelappstarter.util
-
-object Constants {
-
-    const val BASE_URL = "http://www.boredapi.com/api/"
-}
