+++ app/src/main/res/drawable/color_gradient_violet.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <gradient
+        android:startColor="#8A36FF"
+        android:endColor="#C0B2FF"
+        android:angle="45" />
+</shape>
