--- app/src/main/res/menu/menu_details.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:id="@+id/favorite"
-        android:title="@string/favorite" />
-
-</menu>
