--- app/src/main/res/menu/menu_nav.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:id="@+id/listCharacterFragment"
-        android:icon="@drawable/ic_home"
-        android:title="Início" />
-
-    <item
-        android:id="@+id/searchCharacterFragment"
-        android:icon="@drawable/ic_search"
-        android:title="Pesquisar" />
-
-    <item
-        android:id="@+id/favoriteCharacterFragment"
-        android:icon="@drawable/ic_favorite_border"
-        android:title="Favoritos" />
-
-</menu>
