+++ app/src/main/res/navigation/nav_graph.xml
@@ -6,39 +6,9 @@
     app:startDestination="@id/listCharacterFragment">
 
     <fragment
-        android:id="@+id/detailsCharacterFragment"
-        android:name="joaquim.lop.io.marvelappstarter.ui.details.DetailsCharacterFragment"
-        android:label="DetailsCharacterFragment"
-        tools:layout="@layout/fragment_details_character">
-        <argument
-            android:name="character"
-            app:argType="joaquim.lop.io.marvelappstarter.data.model.character.CharacterModel" />
-    </fragment>
-    <fragment
-        android:id="@+id/favoriteCharacterFragment"
-        android:name="joaquim.lop.io.marvelappstarter.ui.favorite.FavoriteCharacterFragment"
-        android:label="FavoriteCharacterFragment"
-        tools:layout="@layout/fragment_favorite_character">
-        <action
-            android:id="@+id/action_favoriteCharacterFragment_to_detailsCharacterFragment"
-            app:destination="@id/detailsCharacterFragment" />
-    </fragment>
-    <fragment
         android:id="@+id/listCharacterFragment"
-        android:name="joaquim.lop.io.marvelappstarter.ui.list.ListCharacterFragment"
+        android:name="joaquim.lop.io.bored.ui.list.ListBoredFragment"
         android:label="ListCharacterFragment"
-        tools:layout="@layout/fragment_list_character">
-        <action
-            android:id="@+id/action_listCharacterFragment_to_detailsCharacterFragment"
-            app:destination="@id/detailsCharacterFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/searchCharacterFragment"
-        android:name="joaquim.lop.io.marvelappstarter.ui.search.SearchCharacterFragment"
-        android:label="SearchCharacterFragment"
-        tools:layout="@layout/fragment_search_character">
-        <action
-            android:id="@+id/action_searchCharacterFragment_to_detailsCharacterFragment"
-            app:destination="@id/detailsCharacterFragment" />
-    </fragment>
+        tools:layout="@layout/fragment_list_bored" />
+
 </navigation>
