+++ app/src/main/res/raw/splash.json
@@ -1,0 +1,1 @@
+{"v":"5.1.20","fr":29.9700012207031,"ip":0,"op":55.0000022401959,"w":100,"h":100,"nm":"flac_webview","ddd":0,"assets":[],"layers":[{"ddd":0,"ind":1,"ty":4,"nm":"flac_webview 윤곽선","sr":1,"ks":{"o":{"a":0,"k":100,"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[50,50,0],"ix":2},"a":{"a":0,"k":[50,50,0],"ix":1},"s":{"a":0,"k":[100,100,100],"ix":6}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[1.781,0],[1.596,-0.946],[0,0],[0,-3.783],[0,0],[-3.192,-1.892],[0,0],[-1.781,0],[-1.596,0.946],[0,0],[0,3.785],[0,0],[3.192,1.892],[0,0]],"o":[[-1.781,0],[0,0],[-3.192,1.892],[0,0],[0,3.785],[0,0],[1.596,0.946],[1.782,0],[0,0],[3.192,-1.892],[0,0],[0,-3.783],[0,0],[-1.596,-0.946]],"v":[[0,-26.207],[-5.159,-24.788],[-18.337,-16.98],[-23.495,-7.809],[-23.495,7.809],[-18.337,16.981],[-5.159,24.789],[0,26.207],[5.159,24.789],[18.337,16.981],[23.495,7.809],[23.495,-7.809],[18.337,-16.98],[5.159,-24.788]],"c":true},"ix":2},"nm":"패스 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"st","c":{"a":0,"k":[0.125,0.713999968884,1,1],"ix":3},"o":{"a":0,"k":100,"ix":4},"w":{"a":0,"k":2,"ix":5},"lc":1,"lj":1,"ml":10,"nm":"선 1","mn":"ADBE Vector Graphic - Stroke","hd":false},{"ty":"tr","p":{"a":1,"k":[{"i":{"x":0.667,"y":1},"o":{"x":0.333,"y":0},"n":"0p667_1_0p333_0","t":0,"s":[49.988,30.051],"e":[49.988,46.301],"to":[0,2.70833325386047],"ti":[0,-2.08333325386047]},{"i":{"x":0.667,"y":1},"o":{"x":0.333,"y":0},"n":"0p667_1_0p333_0","t":27,"s":[49.988,46.301],"e":[49.988,42.551],"to":[0,2.08333325386047],"ti":[0,0.625]},{"t":41.0000016699642}],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":1,"k":[{"i":{"x":[0.667],"y":[1]},"o":{"x":[0.333],"y":[0]},"n":["0p667_1_0p333_0"],"t":0,"s":[0],"e":[62.069]},{"i":{"x":[0.667],"y":[1]},"o":{"x":[0.333],"y":[0]},"n":["0p667_1_0p333_0"],"t":15,"s":[62.069],"e":[0]},{"t":41.0000016699642}],"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"변형"}],"nm":"1","np":2,"cix":2,"ix":1,"mn":"ADBE Vector Group","hd":false},{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[1.781,0],[1.596,-0.946],[0,0],[0,-3.783],[0,0],[-3.192,-1.892],[0,0],[-1.781,0],[-1.596,0.946],[0,0],[0,3.785],[0,0],[3.192,1.892],[0,0]],"o":[[-1.781,0],[0,0],[-3.192,1.892],[0,0],[0,3.785],[0,0],[1.596,0.946],[1.782,0],[0,0],[3.192,-1.892],[0,0],[0,-3.783],[0,0],[-1.596,-0.946]],"v":[[0,-26.207],[-5.159,-24.788],[-18.337,-16.98],[-23.495,-7.809],[-23.495,7.809],[-18.337,16.981],[-5.159,24.789],[0,26.207],[5.159,24.789],[18.337,16.981],[23.495,7.809],[23.495,-7.809],[18.337,-16.98],[5.159,-24.788]],"c":true},"ix":2},"nm":"패스 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"st","c":{"a":0,"k":[0.455000005984,0.152999997606,0.74900004069,1],"ix":3},"o":{"a":0,"k":100,"ix":4},"w":{"a":0,"k":2,"ix":5},"lc":1,"lj":1,"ml":10,"nm":"선 1","mn":"ADBE Vector Graphic - Stroke","hd":false},{"ty":"tr","p":{"a":1,"k":[{"i":{"x":0.667,"y":1},"o":{"x":0.333,"y":0},"n":"0p667_1_0p333_0","t":4,"s":[50,36.782],"e":[50,57.032],"to":[0,3.375],"ti":[0,-2.625]},{"i":{"x":0.667,"y":1},"o":{"x":0.167,"y":0},"n":"0p667_1_0p167_0","t":31,"s":[50,57.032],"e":[50,52.532],"to":[0,2.625],"ti":[0,0.75]},{"t":45.0000018328876}],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":1,"k":[{"i":{"x":[0.667],"y":[1]},"o":{"x":[0.333],"y":[0]},"n":["0p667_1_0p333_0"],"t":4,"s":[0],"e":[100]},{"i":{"x":[0.667],"y":[1]},"o":{"x":[0.333],"y":[0]},"n":["0p667_1_0p333_0"],"t":19,"s":[100],"e":[0]},{"t":45.0000018328876}],"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"변형"}],"nm":"2","np":2,"cix":2,"ix":2,"mn":"ADBE Vector Group","hd":false},{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[1.781,0],[1.596,-0.946],[0,0],[0,-3.783],[0,0],[-3.192,-1.892],[0,0],[-1.781,0],[-1.596,0.946],[0,0],[0,3.785],[0,0],[3.192,1.892],[0,0]],"o":[[-1.781,0],[0,0],[-3.192,1.892],[0,0],[0,3.785],[0,0],[1.596,0.946],[1.782,0],[0,0],[3.192,-1.892],[0,0],[0,-3.783],[0,0],[-1.596,-0.946]],"v":[[0,-26.207],[-5.159,-24.788],[-18.337,-16.98],[-23.495,-7.809],[-23.495,7.809],[-18.337,16.981],[-5.159,24.789],[0,26.207],[5.159,24.789],[18.337,16.981],[23.495,7.809],[23.495,-7.809],[18.337,-16.98],[5.159,-24.788]],"c":true},"ix":2},"nm":"패스 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"st","c":{"a":0,"k":[0.685999971278,0.180000005984,0.980000035903,1],"ix":3},"o":{"a":0,"k":100,"ix":4},"w":{"a":0,"k":2,"ix":5},"lc":1,"lj":1,"ml":10,"nm":"선 1","mn":"ADBE Vector Graphic - Stroke","hd":false},{"ty":"tr","p":{"a":1,"k":[{"i":{"x":0.667,"y":1},"o":{"x":0.333,"y":0},"n":"0p667_1_0p333_0","t":8,"s":[50,44.032],"e":[50,67.782],"to":[0,3.95833325386047],"ti":[0,-2.91666674613953]},{"i":{"x":0.667,"y":1},"o":{"x":0.167,"y":0},"n":"0p667_1_0p167_0","t":35,"s":[50,67.782],"e":[50,61.532],"to":[0,2.91666674613953],"ti":[0,1.04166662693024]},{"t":48.0000019550801}],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":1,"k":[{"i":{"x":[0.667],"y":[1]},"o":{"x":[0.333],"y":[0]},"n":["0p667_1_0p333_0"],"t":8,"s":[0],"e":[100]},{"i":{"x":[0.667],"y":[1]},"o":{"x":[0.333],"y":[0]},"n":["0p667_1_0p333_0"],"t":23,"s":[100],"e":[0]},{"t":48.0000019550801}],"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"변형"}],"nm":"3","np":2,"cix":2,"ix":3,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":55.0000022401959,"st":0,"bm":0}],"markers":[]}
