+++ app/src/main/res/values/colors.xml
@@ -9,4 +9,5 @@
     <color name="white">#FFFFFFFF</color>
     <color name="red">#E91E63</color>
     <color name="light_grey">#B5B5B5</color>
+    <color name="purple">#6532FF</color>
 </resources>
